+++ backend/src/main/java/com/zeta/backend/repository/TicketRepository.java
@@ -14,5 +14,5 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     Optional<Ticket> findTicketsById(Integer ticketId);
 
     // Finds all tickets raised by a specific userId and returns a list of Ticket objects belonging to that user.
-    List<Ticket> findByUserId(Integer userId);
+    List<Ticket> findByUserId(Long userId);
 }
